+++ mod/variables.lua
@@ -470,7 +470,7 @@ function __DebugAdapter.variables(variablesReference,seq,filter,start,count)
               -- skip ahead some names? limit them? vscode does not currently ask for limited names
             end
             for k,v in f,t,firstk do
-              if filter == "indexed" and (type(k) ~= "number" or k > maxindex) then
+              if filter == "indexed" and ((type(k) ~= "number") or (k > maxindex) or (k == 0) or (k % 1 ~= 0)) then
                 break
               end
               local evalName